refactor: move shared backend types into dedicated files
