+++ backend/src/routes/exchange-api-keys.ts
@@ -14,10 +14,8 @@ import { encryptKey } from '../util/crypto.js';
 import { errorResponse, ERROR_MESSAGES } from '../util/error-messages.js';
 import { REDACTED_KEY } from './_shared/constants.js';
 import { getValidatedUserId, userPreHandlers } from './_shared/guards.js';
-import {
-  disableUserWorkflowsByExchangeKey,
-  type DisableWorkflowsSummary,
-} from '../workflows/disable.js';
+import { disableUserWorkflowsByExchangeKey } from '../workflows/disable.js';
+import type { DisableWorkflowsSummary } from '../workflows/disable.types.js';
 import {parseBody} from "./_shared/validation.js";
 
 interface ExchangeKeyBody {
