+++ backend/src/util/error-messages.ts
@@ -1,3 +1,5 @@
+import type { ErrorResponse } from './error-messages.types.js';
+
 export const ERROR_MESSAGES = {
   forbidden: 'forbidden',
   notFound: 'not found',
@@ -7,10 +9,8 @@ export function lengthMessage(field: string, max: number) {
   return `${field} too long (max ${max})`;
 }
 
-export interface ErrorResponse {
-  error: string;
-}
-
 export function errorResponse(message: string): ErrorResponse {
   return { error: message };
 }
+
+export type { ErrorResponse } from './error-messages.types.js';
