+++ backend/src/util/error-messages.types.ts
@@ -1,0 +1,3 @@
+export interface ErrorResponse {
+  error: string;
+}
