+++ backend/src/util/parse-exec-log.types.ts
@@ -1,0 +1,20 @@
+export interface ExecOrder {
+  pair: string;
+  token: string;
+  side: string;
+  quantity: number;
+  limitPrice?: number;
+  basePrice?: number;
+  maxPriceDivergencePct?: number;
+}
+
+export interface ParsedExecLogResponse {
+  orders: ExecOrder[];
+  shortReport: string;
+}
+
+export interface ParsedExecLog {
+  text: string;
+  response?: ParsedExecLogResponse;
+  error?: Record<string, unknown>;
+}
