+++ backend/src/util/tokens.types.ts
@@ -1,0 +1,4 @@
+export interface TokenTags {
+  symbol: string;
+  tags: string[];
+}
