+++ backend/src/workflows/disable.types.ts
@@ -1,0 +1,4 @@
+export interface DisableWorkflowsSummary {
+  disabledWorkflowIds: string[];
+  unscheduledWorkflowIds: string[];
+}
