+++ backend/src/workflows/portfolio-review.ts
@@ -4,11 +4,8 @@ import {
   getActivePortfolioWorkflowsByInterval,
   type ActivePortfolioWorkflow,
 } from '../repos/portfolio-workflows.js';
-import {
-  run as runMainTrader,
-  collectPromptData,
-  type MainTraderDecision,
-} from '../agents/main-trader.js';
+import { run as runMainTrader, collectPromptData } from '../agents/main-trader.js';
+import type { MainTraderDecision } from '../agents/main-trader.types.js';
 import { runNewsAnalyst } from '../agents/news-analyst.js';
 import { runTechnicalAnalyst } from '../agents/technical-analyst.js';
 import { insertReviewRawLog } from '../repos/review-raw-log.js';
